+++ test/kubernetes_test.exs
@@ -81,7 +81,7 @@ defmodule Cluster.Strategy.KubernetesTest do
       end
     end
 
-    test "works with resource version" do
+    test "works with cached resources" do
       use_cassette "kubernetes", custom: true do
         capture_log(fn ->
           start_supervised!({Kubernetes,
@@ -91,7 +91,34 @@ defmodule Cluster.Strategy.KubernetesTest do
                config: [
                  kubernetes_node_basename: "test_basename",
                  kubernetes_selector: "app=test_selector",
-                 kubernetes_resource_version: 0,
+                 kubernetes_use_cached_resources: true,
+                 # If you want to run the test freshly, you'll need to create a DNS Entry
+                 kubernetes_master: "cluster.localhost.",
+                 kubernetes_service_account_path:
+                   Path.join([__DIR__, "fixtures", "kubernetes", "service_account"])
+               ],
+               connect: {Nodes, :connect, [self()]},
+               disconnect: {Nodes, :disconnect, [self()]},
+               list_nodes: {Nodes, :list_nodes, [[]]}
+             }
+           ]})
+
+          assert_receive {:connect, _}, 5_000
+        end)
+      end
+    end
+
+    test "works with no cached resources" do
+      use_cassette "kubernetes", custom: true do
+        capture_log(fn ->
+          start_supervised!({Kubernetes,
+           [
+             %Cluster.Strategy.State{
+               topology: :name,
+               config: [
+                 kubernetes_node_basename: "test_basename",
+                 kubernetes_selector: "app=test_selector",
+                 kubernetes_use_cached_resources: false,
                  # If you want to run the test freshly, you'll need to create a DNS Entry
                  kubernetes_master: "cluster.localhost.",
                  kubernetes_service_account_path:
@@ -228,7 +255,7 @@ defmodule Cluster.Strategy.KubernetesTest do
       end
     end
 
-    test "works with pods and resource version" do
+    test "works with pods and cached resources" do
       use_cassette "kubernetes_pods", custom: true do
         capture_log(fn ->
           start_supervised!({Kubernetes,
@@ -241,7 +268,7 @@ defmodule Cluster.Strategy.KubernetesTest do
                  # If you want to run the test freshly, you'll need to create a DNS Entry
                  kubernetes_master: "cluster.localhost.",
                  kubernetes_ip_lookup_mode: :pods,
-                 kubernetes_resource_version: 0,
+                 kubernetes_use_cached_resources: true,
                  kubernetes_service_account_path:
                    Path.join([__DIR__, "fixtures", "kubernetes", "service_account"])
                ],